Clean up cache tests
