+++ packages/core/src/cache/__test__/cache.spec.ts
@@ -9,9 +9,10 @@ import { createQuery } from '../../query/create_query';
 import { inMemoryCache } from '../adapters/in_memory';
 import { cache } from '../cache';
 import { sha1 } from '../lib/hash';
+import { parseTime } from '../lib/time';
 
 describe('cache', () => {
-  test('use value from cache on second call', async () => {
+  test('use value from cache on second call, revalidate', async () => {
     let index = 1;
     const handler = vi.fn(async (p: void) =>
       setTimeout(1000).then(() => index++)
@@ -26,6 +27,72 @@ describe('cache', () => {
     expect(scope.getState(query.$data)).toEqual(1);
 
     await allSettled(query.reset, { scope });
+
+    // Do not await
+    allSettled(query.start, { scope });
+    // But wait for next tick becuase of async adapter's nature
+    await setTimeout(1);
+
+    // Value from cache
+    expect(scope.getState(query.$data)).toEqual(1);
+    expect(scope.getState(query.$stale)).toBeTruthy();
+
+    await allPrevSettled(scope);
+
+    // After refetch, it's new value
+    expect(scope.getState(query.$data)).toEqual(2);
+    expect(scope.getState(query.$stale)).toBeFalsy();
+
+    expect(handler).toBeCalledTimes(2);
+  });
+
+  test('use value from cache on second call, do not revalidate until staleAfter', async () => {
+    let index = 1;
+    const handler = vi.fn(async (p: void) =>
+      setTimeout(1000).then(() => index++)
+    );
+    const query = withFactory({ fn: () => createQuery({ handler }), sid: '1' });
+
+    cache(query, { adapter: inMemoryCache(), staleAfter: '1h' });
+
+    const scope = fork();
+
+    await allSettled(query.start, { scope });
+    expect(scope.getState(query.$data)).toEqual(1);
+
+    await allSettled(query.reset, { scope });
+
+    // Do not await
+    allSettled(query.start, { scope });
+    // But wait for next tick becuase of async adapter's nature
+    await setTimeout(1);
+
+    // Value from cache
+    expect(scope.getState(query.$data)).toEqual(1);
+    expect(scope.getState(query.$stale)).toBeFalsy();
+
+    await allPrevSettled(scope);
+
+    // No refetch
+    expect(handler).toBeCalledTimes(1);
+  });
+
+  test('mark value as stale since afterStale', async () => {
+    let index = 1;
+    const handler = vi.fn(async (p: void) =>
+      setTimeout(100).then(() => index++)
+    );
+    const query = withFactory({ fn: () => createQuery({ handler }), sid: '1' });
+
+    cache(query, { adapter: inMemoryCache(), staleAfter: '1sec' });
+
+    const scope = fork();
+
+    await allSettled(query.start, { scope });
+    expect(scope.getState(query.$data)).toEqual(1);
+
+    await allSettled(query.reset, { scope });
+    await setTimeout(parseTime('1sec')); // wait for stale after time
 
     // Do not await
     allSettled(query.start, { scope });
